Weather card
Made the first iteration of the weather card, It dosent have any weather icons as I was not able to run the server on my machine and display the images from OpenWeather. The Card currnetly looks like it's only the bare bones, but once I figure out how to make the card look like it's lifted from the page, I change the colors and add the weather images it will look a lot better.